+++ HISTORY.rst
@@ -21,6 +21,7 @@ Unreleased
 * ALCN-1286: Set purchase with uom from dcfunv with value "1" as "Pièces" for the few cases that were failing
 * ALCN-1223: Load and Migrate Suppliers Payment Terms. Modify Client Payment Term 'Immediate Payment' to '7 days' during migration.
   ALCN-1223: Load and Migrate Suppliers Payment Modes.
+* ALCN-1274: Skip order lines with no product ref as those lines are replaced products
 
 **Features and Improvements**
 * Fix Belgium Chart of taxes. Code and name have been inverted in v10.